voicepro-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
PyMySQL==1.1.0
email-validator==2.1.0
google-cloud-speech==2.21.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
requests==2.31.0
SpeechRecognition==3.10.0
pydub==0.25.1
gunicorn==21.2.0
openai>=1.30.0
